// src/@type-app/data-types.d.ts
export interface IDataTypes {
    ARRAY: any,
    VIRTUAL: any,
    UUIDV4: any,
    UUIDV1: any,
    UUID: any,
    TINYINT: any,
    TIME: any,
    TEXT: any,
    STRING: any,
    SMALLINT: any,
    REAL: any,
    RANGE: any,
    NUMBER: any,
    NOW: any,
    MEDIUMINT: any,
    MACADDR: any,
    JSONTYPE: any,
    JSONB: any,
    INTEGER: any,
    INET: any,
    HSTORE: any,
    GEOMETRY: any,
    GEOGRAPHY: any,
    FLOAT: any,
    ENUM: any,
    DOUBLE: any,
    DECIMAL: any,
    DATEONLY: any,
    DATE: any,
    CITEXT: any,
    CIDR: any,
    CHAR: any,
    BOOLEAN: any,
    BLOB: any,
    BIGINT: any
}